Rename: CurrencyInputField -> FieldCurrencyInput
