+++ src/components/FieldCurrencyInput/FieldCurrencyInput.css
